move several http related util methods
